+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All Notable changes to `laravel-modules` will be documented in this file.
     This allows to have the content of the module under an `src/` folder for example.
 - **BREAKING** Renamed method `enabled` to `isEnabled` in `\Nwidart\Modules\Module`.
 - **BREAKING** Renamed method `disabled` to `isDisabled` in `\Nwidart\Modules\Module`.
-
+- **BREAKING** Removed the `__get` magic method on the `\Nwidart\Modules\Module` class. Use `get()` or `json()->get()` instead.
 
 ## 5.0.1 - 2019-05-11